requirements/base.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM and AI
groq>=0.4.0
sentence-transformers>=2.2.0
spacy>=3.7.0

# Web search and scraping
duckduckgo-search>=4.1.0
httpx>=0.25.0
beautifulsoup4>=4.12.0

# Vector store and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.0

# UI
streamlit>=1.28.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# System monitoring
psutil>=5.9.0

# Async support
asyncio-mqtt>=0.16.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP client
requests>=2.31.0

# JSON processing
orjson>=3.9.0

# Date and time
python-dateutil>=2.8.0

# Security
cryptography>=41.0.0

# File handling
aiofiles>=23.0.0

# New additions
langchain-groq
langchain-community 